fix: eui_canvas_str_width 中 lookup_glyph 返回值 int16_t 溢位
carousel 绘制时通过 eui_canvas_str_width 计算中文字串宽度，
但其 u8g2 路径中 idx 仍为 int16_t/decode_glyph_at 传 uint16_t，
导致大字体偏移量溢位，宽度计算为0，文本居中对齐失败，文字重叠

+++ src/eui_canvas.c
@@ -634,10 +634,10 @@ uint16_t eui_canvas_str_width(const eui_canvas_t *canvas, const char *str)
         const char *s = str;
         while (*s) {
             uint32_t cp = utf8_decode_next(&s);
-            int16_t idx = canvas->font->lookup_glyph(canvas->font, (uint16_t)cp, 0);
+            int32_t idx = canvas->font->lookup_glyph(canvas->font, (uint16_t)cp, 0);
             if (idx >= 0) {
                 u8g2_glyph_t g;
-                if (decode_glyph_at(canvas->font, (uint16_t)idx, &g)) {
+                if (decode_glyph_at(canvas->font, (uint32_t)idx, &g)) {
                     w += g.x_advance;
                 }
             }
